refactor(spf): extract the selection-rule composers into a primitive
Moves `SelectionRule`, `SelectionRuleDeps`, `applyRules`, and `applyConstraints`
out of `behaviors/track-switching.ts` into `primitives/selection-rules.ts`. Pure
move: same definitions, same tests, same behavior.

The point is letting both track-selection behaviors share rules. `selectVideoTrack`
exists specifically to tree-shake the ABR path out — that is its documented
purpose and what the background-video variant is built on — so importing a
composer from `track-switching.ts` would drag the bandwidth estimator and quality
selection back in. The four primitives have no dependencies at all, being pure
generics over a candidate list, so from their own module either side reaches them
freely.

`track-switching.ts` re-exports both types and both functions, so nothing typing
against its rules needs a second import. Their tests move alongside them into
`primitives/tests/selection-rules.test.ts`, where they read as what they are:
pure composer tests with no signals, behavior, or engine involved.
